fix(mypy): utils 全量清理至基线 317 — 泛型参数补齐/Optional 守卫/裸容器注解 + risk 批次5
覆盖 utils/ 97 文件: ① 裸 dict/list 注解补 [str,Any] 等参数; ② asyncio.Queue/Task/Callable 补泛型; ③ 自定义守卫不窄化 Optional 处改 assert; ④ 门禁 --no-pretty + mypy_baseline_v9.2.txt 重冻结 830→317; ⑤ 附带 09-08 少量防御加固(glm5_decision_engine 路由降级/safe_math 分母守卫)。

+++ utils/iv_surface_deep_hedge.py
@@ -171,7 +171,7 @@ class SecondOrderGreeks:
         )
         d2 = d1 - vol * np.sqrt(maturity)
         n_prime_d1 = np.exp(-0.5 * d1**2) / np.sqrt(2 * np.pi)
-        return -n_prime_d1 * d2 / vol
+        return float(-n_prime_d1 * d2 / vol)
 
     @staticmethod
     def volga(
@@ -189,7 +189,7 @@ class SecondOrderGreeks:
         d2 = d1 - vol * np.sqrt(maturity)
         n_prime_d1 = np.exp(-0.5 * d1**2) / np.sqrt(2 * np.pi)
         vega = spot * n_prime_d1 * np.sqrt(maturity)
-        return vega * d1 * d2 / vol
+        return float(vega * d1 * d2 / vol)
 
     @staticmethod
     def all_vol_greeks(
@@ -392,7 +392,7 @@ def main() -> None:
         )
 
     spot = 100.0
-    strikes = [90, 95, 100, 105, 110]
+    strikes = [90.0, 95.0, 100.0, 105.0, 110.0]
     maturities = [30 / 365] * 5
     ivs = [0.25, 0.22, 0.20, 0.19, 0.18]
